XVNC port range increase + tips/tricks
The port range in this document is no longer correct - at present, Foreman will try to connect as high as 5999.  Console was failing all over the place in my environment until I made this change.
Also, rather than keeping a copy of the firewall rule local to every system, it's more efficient to store it on shared storage and symlink.  This allows for quick rule updates and easy rule refreshing.