Foreman handbook
This commit intends to be a guide of sorts for Foreman developers on
do's and dont's about Foreman. It explains things new developers might
not be aware of, as why plugins exist, how do we handle API changes, and
so forth.

The main idea is for this to become our people-sourced guideline and
that we can refer to it to solve conflicts, onboard people in the
community, and overall lay down what we want to do.

The way I'm making it is by taking all points here that are relevant:
http://s3.amazonaws.com/marquee-test-akiaisur2rgicbmpehea/KMfujN6vTESm1zhB2jCW_Screen%20Shot%202015-07-22%20at%204.16.55%20PM.png