Revamp Community tab and put Blog on main bar
From the view of someone new to the project, and wanting to know more
about what to watch, how to get help, or how to contribute, I think it
makes sense to have all support / media / code & non-code contributions
on a single page.

Also, the blog moves to the top bar now that it's in MD format and starting to
be in better shape.