Add case studies to media page
Case studies seem to me like the most relevant videos for people
wondering whether to use Foreman or not, more than talks, demos or
deep dives, hence why I put them on top.

I use a youtube embed playlist for this, which I suggest we use for the
rest of the videos, or at least for community demos and deep dives. It's
a rather painless way to get this page auto updated without any updates
to this repository.

The only reason why I haven't done deep dives and community demos in a
playlist too is because our playlists in the Foreman channel only accept
videos within that channel. We would have to disable (Set as official
series for this playlist) on the channel to allow videos from outside.