Add events system to website
* Adds FullCalendar and Moment JS/CSS assets
* Add YAML event data source in _data
* Add Event helpers in a Ruby plugin, in _plugins
* Render /events with a calendar and event list, with links between
* Add icalendar gem to generate ICS file at /events/all.ics

Calendar should render correctly in browser-local timezone as all events
are converted to UTC internally.  ICS file likewise is in UTC.  Links on
the Docs and Community pages.

Event yaml is mostly optional, only title and start date and/or time are
required. If not specified, end is assumed to be 1 day or 1 hour later,
depending on the format of start.