Add safety advise about running foreman on existing systems
This commit adds some text to the quickstart guide and the
install instructions itself, that foreman-installer should run on a
clean, single-purpose system and that it's not advisable to run it
on existing systems. The reason is that installer changes configuration
of several components (and even removes some) which may be surprising
for new users.

Closes GH-812