Replicate puppet kick behaviour
Expectation of the SSH provider is to provide similar functionality to `puppet kick`. Without  `--no-daemonize` this command will start the background daemon, which will then trigger a run every 30 minutes by default. If the daemon is already running this will fail to execute as a second daemon cannot be started.

With `--no-daemonize` the command runs puppet agent once and then exits.