Modify the katello upgrade steps.
Customers found issues when not rebooting the machine. The team discussed
that making the reboot step the last step would allow it to remain optional
and ensure that the testing group is verifying a single upgrade process.
This docs change brings upstream and downstream upgrade steps in line
with that discussion.