Rewrite upgrade steps using an OS selector
This copies the OS selector from the quickstart. This makes the upgrade
instructions shorter since you don't need to skip over irrelevant OSes.
For users this is less confusing.

It also removes the foreman-maintain reference. This can be done by
moving the cleanup step to a point where the services should be stopped
anyway. Avoiding a start-stop cycle makes the overall process shorter.

The noop command removes the redundant --dont-save-answers. --noop
already implies that.