Fixes #20977 - Document that rubygem-rubyipmi might cause issues
Unfortunately, there isn't still any version of rubyipmi containing a
fix for what's described in #20977 [0]. We've hit this issue in
production and having this described in the documentation would have
saved us some time debugging.

I think that given there hasn't been a release upstream since 2015 and
that the bug is still present in the wild (for instance, the version
of rubyipmi shipped in the Foreman repositories is affected) it's now
time to add something to the manual.

[0] https://projects.theforeman.org/issues/20977