Reflect TLS 1.1 disabling in Foreman Proxy 2.1+
Since Foreman Proxy 2.1 TLS 1.1 is disabled out of the box. This makes
the tls_disabled_versions setting a bit pointless. It can still be used
to disable TLS 1.2, but there's little reason to. It's still useful to
keep the section in to explain the TLS versions disabled by default. See
https://projects.theforeman.org/issues/29252 for more details.