Use a git mailmap to correct contributor names
Git can natively map names. This keeps a mailmap in this repository and
copies it in the (temporarily) cloned repositories if it doesn't exist.
Then it relies on git to do the right thing.

The mailmap is not a complete translation of the existing mapping since
it also contains people who haven't contributed in a long time. It also
adds a few missing items.