Add GitHub Action to deploy
Today we use GitHub Actions to build and deploy using Jenkins. This
implements a deployment using GitHub Actions so we use a single
environment.

The design is to build the website just like we do in CI, but then set
up SSH for rsync followed by the actual rsync.

It introduces 3 secrets that are intended to be set as environment
secrets:

* `SSH_PRIVATE_KEY`
* `SSH_KNOWN_HOSTS`
* `DEPLOY_TARGET`

Link: https://docs.github.com/en/actions/use-cases-and-examples/deploying/deploying-with-github-actions